Update Hit.__init__ signature
Hit.__init__ now must be initialized with an iterable returning at least
one HSP object. This is done to reduce setting Hit ID and description in
too many places. They are now attributes of Hit objects that depends on
the HSPs within the Hit.

Conceptually, this change also made the Hit model closer to the
real-world Hits, where there has to be at least one HSP for a database
entry to be declared a Hit.