Update exonerate-text line splitter regex
This update prevents failure when parsing exonerate results whose
introns are all in the hit sequence. Previously, the regex can only
split the lines with e.g. '  >>>> Target Intron 1 >>>>  ' present. It
can now split with e.g. 'gt.........................ag', which is what
the sequence opposite of an intron shows in the output file.